pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumai-otel-genai"
version = "0.1.0"
description = "Standard GenAI telemetry instrumentation for OpenTelemetry"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "AumAI Contributors", email = "dev@aumai.dev" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-api>=1.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "hypothesis>=6.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/aumai/aumai-otel-genai"
Documentation = "https://docs.aumai.dev/aumai-otel-genai"
Repository = "https://github.com/aumai/aumai-otel-genai"
Issues = "https://github.com/aumai/aumai-otel-genai/issues"

[project.scripts]
aumai-otel-genai = "aumai_otel_genai.cli:main"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "ANN", "S", "B"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
